Refactored the tasks in multiple modules.
